feat: Add update endpoint for admin
This commit adds an update endpoint for the admin user in the user controller. The endpoint allows updating the admin's information based on their DNI. It also includes validation using the userPartialValidator from the userSchema. If the admin is not found, a 404 error is returned. The successful update returns a 200 status code with a success message.