feat: Add date field to venta table
The code changes include adding a new column `fecha` of type `DATE` to the `venta` table in the database. This column will store the date of the sale transaction.

This commit message follows the established convention of using a prefix (`feat`) to indicate the type of change and provides a clear and concise description of the added functionality.