feat: Add fecha column to Venta table
The code changes include adding a new column `fecha` to the `Venta` table in the database. This column will store the date of the sale transaction.

This commit message follows the established convention of using a prefix (`feat`) to indicate the type of change and provides a clear and concise description of the added functionality.